+++ client/src/book-tocs.ts
@@ -1,7 +1,7 @@
 import { EventEmitter, TreeItemCollapsibleState, Uri, TreeDataProvider } from 'vscode'
 
-import { BookToc, ClientTocNode, BookRootNode, TocNodeKind } from '../../common/src/toc-tree'
-import { TocItemIcon } from './toc-trees'
+import { BookToc, ClientTocNode, BookRootNode, TocNodeKind } from '../../common/src/toc'
+import { TocItemIcon } from './toc-trees-provider'
 
 export type BookOrTocNode = BookToc | ClientTocNode
