+++ client/src/toc-trees-provider.ts
@@ -1,5 +1,5 @@
 import vscode, { ThemeIcon } from 'vscode'
-import { BookRootNode, ClientTocNode, TocNodeKind } from '../../common/src/toc-tree'
+import { BookRootNode, ClientTocNode, TocNodeKind } from '../../common/src/toc'
 import { TocsTreeProvider, BookOrTocNode } from './book-tocs'
 import { ExtensionHostContext } from './panel'
