+++ server/src/model-manager.spec.ts
@@ -14,7 +14,7 @@ import { Job, JobRunner } from './job-runner'
 import { PageInfo, pageMaker } from './model/page.spec'
 
 import { PageNode } from './model/page'
-import { TocModification, TocModificationKind, TocNodeKind } from '../../common/src/toc-tree'
+import { TocModification, TocModificationKind, TocNodeKind } from '../../common/src/toc'
 import { BooksAndOrphans, DiagnosticSource } from '../../common/src/requests'
 import { bookMaker } from './model/book.spec'
 import { bundleMaker } from './model/bundle.spec'
